Moved the old google crawler to utils.
